// ebk_front/home/bible/app/lib/config/interface.ts
export interface Eglise {
  createdAt: Date;
  updatedAt: Date;
  deletedAt: null;
  id_eglise: number;
  nom_eglise: string;
  username_eglise: string;
  photo_eglise: string;
  couverture_eglise: string;
  sigle_eglise: string;
  adresse_eglise: string;
  ville_eglise: string;
  pays_eglise: string;
  nombrefidel_eglise: string;
  status_eglise: string;
  payement_eglise: boolean;
  programme?: Programme[];
  word_of_day_eglise: string;
  about_eglise: string;
  localisation_eglise: string[];
}
// ...
export interface Programme {
  createdAt: Date;
  updatedAt: Date;
  deletedAt: null;
  id: number;
  titre: string;
  sousProgramme: SousProgramme[];
}
// ...
export interface Content {
  id: number;
  book: string;
  chapter: string;
  verse: string | null;
}

export interface ContentDayPlan {
  createdAt: string; // Vous pouvez utiliser `Date` si vous voulez manipuler des objets date en JavaScript.
  updatedAt: string; // Idem pour `updatedAt`.
  deletedAt: string | null; // Idem pour `deletedAt`, ou laisser en `null` si non utilisé.
  id: number;
  title: string;
  devotion: string;
  day: number;
  contents: Content[];
}

export interface ItemBiblePlanLecture {
  createdAt: string;
  updatedAt: string;
  deletedAt: string | null;
  id: number;
  title: string;
  description: string;
  categorie: string;
  picture: string;
  number_days: number;
  share: any[];
  eglise: Eglise;
}